Unit test for cache hits.
This change assumes it is a desirable feature of TORAX currently that we don't have any extra compilation occuring when running a simulation for the second time. In this case this is a useful test to add as we have silently added in cache misses to TORAX recently and it is easy to mistakenly add them back in, in future.

It broadly checks if there is any tracing or compilation going on for a second simulation run.

It is prone to internal JAX changing but we will know in either case if it does as the test should fail.

PiperOrigin-RevId: 743889197