Preload ToricNN at config construction time.
We can load and validate the ToricNN at config construction time. Advantages:
- catch any configuration for model loading, errors at config construction time
- load the model once for a given `ToraxConfig`.

As a follow up we should discuss in the short term how to embed the model path such that it's consistent with qlknn, perhaps in the pydantic config where it can be immediately validated rather than through this env var till the longer term solution of surrogate dependencies lands.

PiperOrigin-RevId: 747359397