Output restructure: Rename output variables in core profiles.
Note this change does not rename the names in the `CoreProfiles` object. This can be done at a later stage as it will be reflected internally and also in general it is useful to have the output decoupled from our internals.

This change makes part of a larger change in TORAX output naming. Specifically here we:
- remove _face from names so that variable names are not tied to TORAX grids (instead this information is available in the coordinates for a given field)
- remove short abbreviations like tot->total
- rename other core profiles to be clearer.

PiperOrigin-RevId: 751419673