Refactor plot config to load from absolute paths.
A design question: The current design is that `plot_config=None` uses the default plot config. An alternative is to allow `plot_config` to either take a path, or one of the Torax presets, `'default_plot_config'`, `'global_params_plot_config'`. A similar design question exists for `'config'` (should users be able to give presets? This simplifies examples).

PiperOrigin-RevId: 756213810